Move directory for <Checkbox> and <RadioButton> components
They are composition object rather than block elements.
