+++ src/block-kit/composition/Checkbox.ts
@@ -1,5 +1,5 @@
 import { MrkdwnElement } from '@slack/types'
-import { mrkdwn, mrkdwnForOption } from '../composition/Mrkdwn'
+import { mrkdwn, mrkdwnForOption } from './Mrkdwn'
 import { JSXSlack, createComponent } from '../../jsx'
 
 export const checkboxCheckedSymbol = Symbol('jsx-slack-checkbox-checked')
